migrate more tests
