+++ tests/unit/test__job_helpers.py
@@ -15,7 +15,6 @@
 from typing import Any, Dict, Optional
 from unittest import mock
 
-import freezegun
 import google.api_core.exceptions
 from google.api_core import retry as retries
 import pytest
@@ -448,68 +447,6 @@ def test_query_and_wait_uses_jobs_insert():
         },
         timeout=None,
     )
-
-
-def test_query_and_wait_retries_job(global_time_lock):
-    freezegun.freeze_time(auto_tick_seconds=100)
-    client = mock.create_autospec(Client)
-    client._call_api.__name__ = "_call_api"
-    client._call_api.__qualname__ = "Client._call_api"
-    client._call_api.__annotations__ = {}
-    client._call_api.__type_params__ = ()
-    client._call_api.side_effect = (
-        google.api_core.exceptions.BadGateway("retry me"),
-        google.api_core.exceptions.InternalServerError("job_retry me"),
-        google.api_core.exceptions.BadGateway("retry me"),
-        {
-            "jobReference": {
-                "projectId": "response-project",
-                "jobId": "abc",
-                "location": "response-location",
-            },
-            "jobComplete": True,
-            "schema": {
-                "fields": [
-                    {"name": "full_name", "type": "STRING", "mode": "REQUIRED"},
-                    {"name": "age", "type": "INT64", "mode": "NULLABLE"},
-                ],
-            },
-            "rows": [
-                {"f": [{"v": "Whillma Phlyntstone"}, {"v": "27"}]},
-                {"f": [{"v": "Bhetty Rhubble"}, {"v": "28"}]},
-                {"f": [{"v": "Phred Phlyntstone"}, {"v": "32"}]},
-                {"f": [{"v": "Bharney Rhubble"}, {"v": "33"}]},
-            ],
-        },
-    )
-    rows = _job_helpers.query_and_wait(
-        client,
-        query="SELECT 1",
-        location="request-location",
-        project="request-project",
-        job_config=None,
-        page_size=None,
-        max_results=None,
-        retry=retries.Retry(
-            lambda exc: isinstance(exc, google.api_core.exceptions.BadGateway),
-            multiplier=1.0,
-        ).with_deadline(
-            200.0
-        ),  # Since auto_tick_seconds is 100, we should get at least 1 retry.
-        job_retry=retries.Retry(
-            lambda exc: isinstance(exc, google.api_core.exceptions.InternalServerError),
-            multiplier=1.0,
-        ).with_deadline(600.0),
-    )
-    assert len(list(rows)) == 4
-
-    # For this code path, where the query has finished immediately, we should
-    # only be calling the jobs.query API and no other request path.
-    request_path = "/projects/request-project/queries"
-    for call in client._call_api.call_args_list:
-        _, kwargs = call
-        assert kwargs["method"] == "POST"
-        assert kwargs["path"] == request_path
 
 
 def test_query_and_wait_sets_job_creation_mode():
